remove all old property list code and react-table packages
