--- client/src/styles/PropertiesListFilter.scss
@@ -1,263 +1,0 @@
-@import "_vars.scss";
-@import "_scrollbar.scss";
-@import "_button.scss";
-
-$table-border-light: 1px solid $gray-light;
-$table-border-dark: 1px solid $dark;
-
-.PropertiesList {
-  height: 100%;
-  width: 100%;
-  // These next 2 items ensure that the fade effect from the "::after"
-  // component doesn't overflow past the edge of the PropertiesList container
-  position: relative;
-  overflow: auto;
-
-  #multiselectContainerReact {
-    ul {
-      list-style-type: none;
-    }
-    .chip {
-      color: $wowza-dark;
-    }
-    .close-button {
-      color: $wowza-dark;
-      padding-left: 0.5rem;
-      font-style: normal;
-      height: 100%;
-      cursor: pointer;
-      &:hover {
-        opacity: 0.7;
-      }
-    }
-  }
-
-  table {
-    border-top: none;
-    border-left: none;
-    border-color: $gray-dark;
-    border-collapse: collapse;
-
-    .arrow-icon {
-      display: inline-block;
-      margin-left: 0.5rem;
-      span {
-        // This small adjustment aligns second arrow in icon with first
-        display: inline-block;
-        transform: translateY(0.5px);
-      }
-    }
-
-    .col-address {
-      position: sticky;
-      left: 0px;
-      z-index: 40;
-      background-color: inherit;
-      box-shadow: inset -2px 0 2px -1px #ccc !important;
-    }
-
-    tr {
-      padding: 0px 5px;
-    }
-
-    td {
-      text-align: center;
-      text-overflow: ellipsis;
-      overflow: hidden;
-      transition: 0.3s ease;
-      transition-property: width, min-width, padding, opacity;
-    }
-
-    th,
-    td {
-      border-right: $table-border-light;
-    }
-
-    thead {
-      max-height: 30px;
-
-      tr {
-        height: 30px;
-      }
-
-      tr:first-child th {
-        border-color: $gray-light;
-        border-bottom: $table-border-light;
-        font-weight: bold;
-        color: #000;
-        background-color: #f7f7f7;
-      }
-
-      tr:last-child th {
-        background-color: #fff;
-        border-bottom: $table-border-dark;
-        box-shadow: none;
-        vertical-align: top;
-      }
-
-      th {
-        padding: 7px 5px;
-      }
-    }
-
-    tbody {
-      tr {
-        border-bottom: $table-border-light;
-
-        &.row-odd {
-          background-color: #f7f7f7;
-          &:hover {
-            background-color: #f2f2f2ff;
-          }
-        }
-        &.row-even {
-          background-color: #fff;
-          &:hover {
-            background-color: #f0f0f0;
-          }
-        }
-      }
-
-      td {
-        &.col-address {
-          text-align: left;
-          padding: 0.75rem;
-        }
-        &.col-address,
-        &.col-recentcomplaintsbytype,
-        &.col-ownernames,
-        &.col-yearstartedj51,
-        &.col-yearstarted421a,
-        &.col-detail {
-          white-space: unset;
-        }
-      }
-    }
-  }
-
-  .pagination {
-    z-index: 1;
-    display: flex;
-    justify-content: space-between;
-    align-items: stretch;
-    flex-wrap: wrap;
-    padding: 3px;
-
-    div {
-      display: block;
-      flex: 1 1;
-      text-align: center;
-    }
-
-    .center {
-      flex: 1.5 1;
-      text-align: center;
-      margin-bottom: 0;
-      display: flex;
-      flex-direction: row;
-      flex-wrap: wrap;
-      align-items: center;
-      justify-content: space-around;
-
-      .page-info {
-        display: flex;
-        align-items: center;
-        margin: 3px 10px;
-        white-space: nowrap;
-
-        input {
-          width: 70px;
-          text-align: center;
-          display: inline-block;
-          border: 1px solid rgba(0, 0, 0, 0.1);
-          background: #fff;
-          padding: 5px 7px;
-          margin: 5px 7px;
-          font-size: inherit;
-          border-radius: 3px;
-          font-weight: 400;
-          outline: none;
-        }
-        .total-pages {
-          margin-left: 5px;
-        }
-      }
-    }
-
-    .next,
-    .prev {
-      .page-btn[disabled] {
-        opacity: 0.5;
-        cursor: default;
-      }
-
-      .page-btn {
-        -webkit-appearance: none;
-        -moz-appearance: none;
-        appearance: none;
-        display: block;
-        width: 100%;
-        height: 100%;
-        border: 0;
-        border-radius: 3px;
-        padding: 6px;
-        font-size: 1em;
-        color: rgba(0, 0, 0, 0.6);
-        background: rgba(0, 0, 0, 0.1);
-        transition: all 0.1s ease;
-        cursor: pointer;
-        outline: none;
-      }
-    }
-
-    select {
-      border: 1px solid rgba(0, 0, 0, 0.1);
-      background: #fff;
-      padding: 5px 7px;
-      font-size: inherit;
-      border-radius: 3px;
-      font-weight: 400;
-      outline: none;
-    }
-  }
-
-  // TODO: this doesn't stay on the side of the screen anymore, but gets fixed to the spot on the table that at the edge of the screen at first Add
-
-  // right-side fade out when list is scrollable:
-  // &::after {
-  //   content: "";
-  //   display: block;
-  //   position: absolute;
-  //   pointer-events: none;
-  //   top: 0;
-  //   bottom: 0;
-  //   right: 0;
-  //   width: 50px;
-  //   z-index: 10;
-  //   background: -moz-linear-gradient(
-  //     left,
-  //     rgba(255, 255, 255, 0) 0%,
-  //     rgba(255, 255, 255, 1) 100%
-  //   ); /* FF3.6-15 */
-  //   background: -webkit-linear-gradient(
-  //     left,
-  //     rgba(255, 255, 255, 0) 0%,
-  //     rgba(255, 255, 255, 1) 100%
-  //   ); /* Chrome10-25,Safari5.1-6 */
-  //   background: linear-gradient(
-  //     to right,
-  //     rgba(255, 255, 255, 0) 0%,
-  //     rgba(255, 255, 255, 1) 100%
-  //   ); /* W3C, IE10+, FF16+, Chrome26+, Opera12+, Safari7+ */
-  //   filter: progid:DXImageTransform.Microsoft.gradient( startColorstr='#00ffffff', endColorstr='#ffffff',GradientType=1 ); /* IE6-9 */
-  //   visibility: visible;
-  //   opacity: 1;
-  //   transition: opacity 0.5s linear;
-  // }
-
-  // &.hide-scroll-fade::after {
-  //   visibility: hidden;
-  //   opacity: 0;
-  //   transition: visibility 0s 0.5s, opacity 0.5s linear;
-  // }
-}
